Further cleanup and optimization. Changed multiply/divide operations to shifts where possible, changed the system palette array from using RGB structs to straight 32 bit unsigned integers reducing space used and number of operations when composing pixel colors. Also changed makefile to use O3 optimizations and removed more unnecessary comments.

+++ src/CPU.hpp
@@ -27,7 +27,7 @@ class CPU {
         unsigned char STACK_POP();
         void RESET();
         void IRQ_NMI(uint16_t V);
-        void WAIT(uint8_t N=1);
+        void WAIT(uint16_t N=1);
 
         Cartridge *ROM;
         PPU *P;
